issue#14: Create ConsoleEmailService for dev profile
