+++ src/main/java/com/metalr2/service/email/ConsoleEmailService.java
@@ -1,0 +1,35 @@
+package com.metalr2.service.email;
+
+import com.metalr2.model.email.AbstractEmail;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Profile;
+import org.springframework.stereotype.Service;
+import org.thymeleaf.context.Context;
+import org.thymeleaf.spring5.SpringTemplateEngine;
+
+@Service
+@Slf4j
+@Profile("dev")
+public class ConsoleEmailService implements EmailService {
+
+  private final SpringTemplateEngine templateEngine;
+
+  @Autowired
+  public ConsoleEmailService(SpringTemplateEngine templateEngine) {
+    this.templateEngine = templateEngine;
+  }
+
+  @Override
+  public void sendEmail(AbstractEmail email) {
+    Context context = new Context();
+    context.setVariables(email.getViewModel());
+    String messageAsHtml = templateEngine.process(email.getTemplateName(), context);
+
+    log.debug("From: {}", email.getFrom());
+    log.debug("Recipient: {}", email.getRecipient());
+    log.debug("Subject: {}", email.getSubject());
+    log.debug("Message as Html: {}", messageAsHtml);
+  }
+
+}
